+++ include/mu/tiny/test/Failure.hpp
@@ -48,7 +48,7 @@ public:
   Failure(
       Shell* test,
       const char* file_name,
-      size_t line_number,
+      int line_number,
       const String& the_message
   );
 
@@ -68,7 +68,7 @@ public:
    * @param file_name    Source file of the failing assertion.
    * @param line_number  Line number of the failing assertion.
    */
-  Failure(Shell* test, const char* file_name, size_t line_number);
+  Failure(Shell* test, const char* file_name, int line_number);
   Failure(const Failure&) = default;
   Failure(Failure&& other) noexcept;
   virtual ~Failure() = default;
@@ -80,13 +80,13 @@ public:
   /** @return Test case name without the group prefix. */
   virtual const String& get_test_name_only() const;
   /** @return Line number of the failing assertion. */
-  virtual size_t get_failure_line_number() const;
+  virtual int get_failure_line_number() const;
   /** @return Human-readable failure description. */
   virtual const String& get_message() const;
   /** @return Source file in which the test is defined. */
   virtual const String& get_test_file_name() const;
   /** @return Line number of the TEST() macro for the failing test. */
-  virtual size_t get_test_line_number() const;
+  virtual int get_test_line_number() const;
 
   /**
    * @return true if this failure represents an error (e.g. unexpected
@@ -138,9 +138,9 @@ protected:
   String test_name_;
   String test_name_only_;
   String file_name_;
-  size_t line_number_;
+  int line_number_;
   String test_file_name_;
-  size_t test_line_number_;
+  int test_line_number_;
   String message_;
 
   Failure& operator=(const Failure&);
@@ -163,7 +163,7 @@ public:
   EqualsFailure(
       Shell* test,
       const char* file_name,
-      size_t line_number,
+      int line_number,
       const char* expected,
       const char* actual,
       const String& text
@@ -179,7 +179,7 @@ public:
   EqualsFailure(
       Shell* test,
       const char* file_name,
-      size_t line_number,
+      int line_number,
       const String& expected,
       const String& actual,
       const String& text
@@ -211,7 +211,7 @@ public:
   ApproxEqualFailure(
       Shell* test,
       const char* file_name,
-      size_t line_number,
+      int line_number,
       T expected,
       T actual,
       T threshold,
@@ -249,7 +249,7 @@ public:
   CheckEqualFailure(
       Shell* test,
       const char* file_name,
-      size_t line_number,
+      int line_number,
       const String& expected,
       const String& actual,
       const String& text
@@ -273,7 +273,7 @@ public:
   ComparisonFailure(
       Shell* test,
       const char* file_name,
-      size_t line_number,
+      int line_number,
       const String& check_string,
       const String& comparison_string,
       const String& text
@@ -298,7 +298,7 @@ public:
   ContainsFailure(
       Shell* test,
       const char* file_name,
-      size_t line_number,
+      int line_number,
       const String& expected,
       const String& actual,
       const String& text
@@ -322,7 +322,7 @@ public:
   CheckFailure(
       Shell* test,
       const char* file_name,
-      size_t line_number,
+      int line_number,
       const String& check_string,
       const String& condition_string,
       const String& text_string = ""
@@ -344,7 +344,7 @@ public:
   FailFailure(
       Shell* test,
       const char* file_name,
-      size_t line_number,
+      int line_number,
       const String& message
   );
 };
@@ -370,7 +370,7 @@ public:
   IntMaxEqualFailure(
       Shell* test,
       const char* file_name,
-      size_t line_number,
+      int line_number,
       long long expected,
       long long actual,
       const String& text
@@ -397,7 +397,7 @@ public:
   UintMaxEqualFailure(
       Shell* test,
       const char* file_name,
-      size_t line_number,
+      int line_number,
       unsigned long long expected,
       unsigned long long actual,
       const String& text
@@ -421,7 +421,7 @@ public:
   StringEqualFailure(
       Shell* test,
       const char* file_name,
-      size_t line_number,
+      int line_number,
       const char* expected,
       const char* actual,
       const String& text
@@ -445,7 +445,7 @@ public:
   StringEqualNoCaseFailure(
       Shell* test,
       const char* file_name,
-      size_t line_number,
+      int line_number,
       const char* expected,
       const char* actual,
       const String& text
@@ -470,7 +470,7 @@ public:
   BinaryEqualFailure(
       Shell* test,
       const char* file_name,
-      size_t line_number,
+      int line_number,
       const unsigned char* expected,
       const unsigned char* actual,
       size_t size,
@@ -494,7 +494,7 @@ public:
   FeatureUnsupportedFailure(
       Shell* test,
       const char* file_name,
-      size_t line_number,
+      int line_number,
       const String& feature_name,
       const String& text
   );
