+++ src/test/Installer.cpp
@@ -12,7 +12,7 @@ Installer::Installer(
     const char* group_name,
     const char* test_name,
     const char* file_name,
-    size_t line_number
+    int line_number
 ) noexcept
 {
   shell.set_group_name(group_name);
