+++ tests/src/CommandLineTestRunner.test.cpp
@@ -77,7 +77,7 @@ public:
       const char* group_name,
       const char* test_name,
       const char* file_name,
-      size_t line_number
+      int line_number
   )
     : SkippedShell(group_name, test_name, file_name, line_number)
   {
